add macro to convert fullmove_number and active_color to a ply_count (number of halfmoves made during the whole game which is different than the halfmove_clock). Change the fullmove_number to a usize since it might as well be used to index directly while making moves

+++ chess_engine/src/error.rs
@@ -231,13 +231,13 @@ pub enum GamestateValidityCheckError {
         "Fullmove number: {fullmove_number} should be in range 1..={}",
         MAX_GAME_MOVES
     )]
-    StrictFullmoveNumberNotInRange { fullmove_number: u32 },
+    StrictFullmoveNumberNotInRange { fullmove_number: usize },
 
     #[error(
         "Given halfmove clock: {halfmove_clock}, fullmove number: {fullmove_number} is too small"
     )]
     StrictFullmoveNumberLessThanHalfmoveClockDividedByTwo {
-        fullmove_number: u32,
+        fullmove_number: usize,
         halfmove_clock: u8,
     },
